requirements.txt
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.3.3
gunicorn==19.6.0
Markdown==2.6.6
psycopg2==2.6.1
whitenoise==3.2
